move Channel extension to test-fixtures module
