+++ godtools-tool-parser/src/commonTest/kotlin/org/cru/godtools/tool/state/StateTest.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import org.cru.godtools.tool.internal.receive
+import org.cru.godtools.tool.internal.coroutines.receive
 import org.cru.godtools.tool.internal.runBlockingTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
